feat: UserService 인터페이스 추가
- 기존 UserService는 AppUserService으로 이름 변경
